Move request process and response handling from HTTPAdapter to HTTPRequester and responsehandler
